requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
bcrypt==4.1.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
